Move implementation of CultureConditions to CultureRuleProvider class
